Modificacion total de todas las vistas del proyecto, todo funcionalmente correcto

+++ proyecto_bancario/db/querysDB.py
@@ -7,19 +7,21 @@ def registro_admin():
         db = connection.cursor()
         query = "INSERT INTO cliente (cedula_cliente, nombre_cliente, apellido_cliente, telefono_cliente, direccion_cliente, codigo_ciudad) VALUES ('123','camila','mendez','1234','cra1236','434');"
         db.execute(query)
-        query = "INSERT INTO adm_cuenta(codigo_adm,cedula_cliente,fecha_creacion_cuenta) VALUES ('2409','123','2024-10-16');"
+        query = "INSERT INTO adm_cuenta(codigo_adm,cedula_cliente,fecha_creacion_cuenta) VALUES ('2409','100555','2024-10-16');"
         db.execute(query)
         connection.commit()
     return
 
-
-def obtener_ciudades_por_nombre():
+def obtener_ciudades():
     db = connection.cursor()
-    query = "SELECT nombre_ciudad FROM ciudad"
+    query = "SELECT codigo_ciudad, nombre_ciudad FROM ciudad"  # Incluye el código aquí si lo necesitas
     db.execute(query)
     result = db.fetchall()
+    db.close()
 
-    return result
+    # Devuelve cada ciudad como un diccionario de código y nombre para simplificar su uso en el template
+    return [{'codigo': row[0], 'nombre': row[1]} for row in result]
+
 
 def obtener_ciudad_por_nombre(nombre_ciudad):
     db = connection.cursor()
@@ -90,14 +92,8 @@ def listar_clientes():
 
     return result
 
-def registro_cliente(Cedula, Nombre, Apellido, Telefono, Direccion, ciudad, codigo_ciudad):
+def registro_cliente(Cedula, Nombre, Apellido, Telefono, Direccion, codigo_ciudad):
     db = connection.cursor()
-    query = "SELECT * FROM ciudad WHERE codigo_ciudad = %s AND nombre_ciudad = %s;"
-    db.execute(query, (str(codigo_ciudad), ciudad))
-    result = db.fetchall()
-    if len(result) == 0:
-        query = "INSERT INTO ciudad (codigo_ciudad, nombre_ciudad) values (%s, %s);"
-        result= db.execute(query, (codigo_ciudad, ciudad))
 
     query = "INSERT INTO cliente (cedula_cliente, nombre_cliente, apellido_cliente, telefono_cliente, direccion_cliente, codigo_ciudad) VALUES (%s, %s, %s, %s, %s, %s);"
     values = (Cedula, Nombre, Apellido, Telefono, Direccion, codigo_ciudad)
@@ -109,18 +105,11 @@ def registro_cliente(Cedula, Nombre, Apellido, Telefono, Direccion, ciudad, codi
     
     return False
 
-def actualizar_usuario (cedula,nombre,apellido,telefono,direccion,codigo_ciudad, ciudad):
+def actualizar_usuario (cedula,nombre,apellido,telefono,direccion,ciudad):
     db = connection.cursor()
 
-    query = "SELECT * FROM ciudad WHERE codigo_ciudad = %s AND nombre_ciudad = %s;"
-    db.execute(query, (str(codigo_ciudad), ciudad))
-    result = db.fetchall()
-    if len(result) == 0:
-        query = "INSERT INTO ciudad (codigo_ciudad, nombre_ciudad) values (%s, %s);"
-        result= db.execute(query, (codigo_ciudad, ciudad))
-
     query = "UPDATE cliente SET nombre_cliente = %s, apellido_cliente = %s, telefono_cliente = %s,direccion_cliente = %s, codigo_ciudad=%s WHERE cedula_cliente = %s;" 
-    values = (nombre,apellido,telefono,direccion,codigo_ciudad,cedula)
+    values = (nombre,apellido,telefono,direccion,ciudad,cedula)
     result = db.execute(query,values)
     connection.commit()
     if result == 1:
@@ -129,22 +118,23 @@ def actualizar_usuario (cedula,nombre,apellido,telefono,direccion,codigo_ciudad,
     return False
 
 def eliminar_cliente(cedula_cliente):
-    db = connection.cursor()
-
     codigo_cuenta = obtener_clientes_por_cedula(cedula_cliente)[0][6]
-    print("codigo_cuenta eliminar:", codigo_cuenta)
+
+    db = connection.cursor()
+    print("codigo_cuenta a eliminar:", codigo_cuenta)
     query = "DELETE FROM cliente WHERE cedula_cliente = %s;"
     result = db.execute(query, cedula_cliente)
     connection.commit()
+    print("Cliente elimiado correctamente")
     if result == 1 and codigo_cuenta == None:
         return True
-    else:
-        print("Cliente elimiado correctamente")
+    if codigo_cuenta != None:
+        print("Vamos a eliminar una cuenta %s del cliente %s", codigo_cuenta, cedula_cliente)
         query = "DELETE FROM cuenta WHERE codigo_cuenta = %s;"
         result = db.execute(query, codigo_cuenta)
         connection.commit()
-        print("cuenta elimiado correctamente")
         if result == 1:
+            print("cuenta elimiado correctamente")
             return True
     
     return False
@@ -172,7 +162,7 @@ def registro_cuenta(codigo_cuenta, cedula_cliente, nombre_cuenta, estado, saldo)
             return True 
     return False
 
-def actualizar_cuenta(codigo_cuenta, nombre_cuenta, estado):
+def actualizar_cuenta(nombre_cuenta, estado, codigo_cuenta):
     db = connection.cursor()
     query = "UPDATE cuenta SET nombre_cuenta = %s, estado = %s WHERE codigo_cuenta = %s;" 
     values = (nombre_cuenta, estado, codigo_cuenta)
@@ -183,15 +173,20 @@ def actualizar_cuenta(codigo_cuenta, nombre_cuenta, estado):
            
     return False
 
-def eliminar_cuenta(codigo_cuenta, cedula_cliente):
+def eliminar_cuenta(codigo_cuenta):
+    cuenta_asociada_eliminada = False
     db = connection.cursor()
 
-    query = "UPDATE cliente SET codigo_cuenta = Null WHERE cedula_cliente = %s AND codigo_cuenta = %s;" 
-    values = (cedula_cliente, str(codigo_cuenta))
-    result = db.execute(query,values)
-    connection.commit()
+    cliente_por_cuenta = obtener_clientes_por_codigo_cuenta(codigo_cuenta)
+    if cliente_por_cuenta != {}:
+        query = "UPDATE cliente SET codigo_cuenta = Null WHERE cedula_cliente = %s AND codigo_cuenta = %s;" 
+        values = (str(cliente_por_cuenta[0][0]), str(codigo_cuenta))
+        result = db.execute(query,values)
+        connection.commit()
+        if result == 1:
+            cuenta_asociada_eliminada = True
 
-    if result == 1:
+    if cuenta_asociada_eliminada or cliente_por_cuenta == {}:
         query = "DELETE FROM cuenta WHERE codigo_cuenta = %s;"
         result = db.execute(query, codigo_cuenta)
         connection.commit()
